robot-framework-web/requirements.txt
py==1.8.0
robotframework==3.1.2
robotframework-debuglibrary==1.1.4
robotframework-jsonlibrary==0.3.1
robotframework-lint==1.0
robotframework-pageobjectlibrary==1.0.2
robotframework-seleniumlibrary==4.1.0
selenium==3.141.0